requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
